Update package structure for easier understanding of activities
