+++ app/src/main/AndroidManifest.xml
@@ -49,7 +49,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SignupActivity"
+            android:name=".auth.SignupActivity"
             android:exported="false"
             android:label="@string/title_activity_signup"
             android:theme="@style/Theme.CoachMe" />
@@ -64,12 +64,12 @@
             android:label="@string/title_activity_schedule"
             android:theme="@style/Theme.CoachMe" />
         <activity
-            android:name=".ProfileActivity"
+            android:name=".auth.ProfileActivity"
             android:exported="false"
             android:label="@string/title_activity_edit_profile"
             android:theme="@style/Theme.CoachMe" />
         <activity
-            android:name=".SelectSportsActivity"
+            android:name=".profile.SelectSportsActivity"
             android:exported="true"
             android:label="@string/title_activity_multiselect_list"
             android:theme="@style/Theme.CoachMe" />
