--- app/src/main/java/com/github/sdpcoachme/data/ListItem.kt
@@ -1,9 +1,0 @@
-package com.github.sdpcoachme.data
-
-/**
- * A data class for a list item
- */
-data class ListItem<A>(
-    val element: A,
-    val selected: Boolean
-)
