--- app/src/main/java/com/github/sdpcoachme/data/MapState.kt
@@ -1,8 +1,0 @@
-package com.github.sdpcoachme.data
-
-import com.google.android.gms.maps.model.LatLng
-
-/**
- * Holds the last known user location as LatLng type.
- */
-data class MapState(val lastKnownLocation: LatLng?)
